Convert external negotiation util to internal pkg
Dep insists on pulling in GDDO when end users pull in this package, even
though it's already checked in and vendored. This causes a ton of issues
with the recent breaking changes in google/protobuf, since GDDO pins to
a very old version of proto.

The simplest way to solve this is to turn the small bits of GDDO we use
into an internal package, which removes dep from the equation
completely.